Adds enableLinks property to 'utfallSvar' components
This commit introduces the `enableLinks` property to several custom element renderers.

The `enableLinks` flag is now passed to `CustomElementHtmlAttributes` to control whether links should be enabled within the rendered elements. This applies to the "beskrivelse", "status", "tema", and "kommentar" elements, as well as the "utfallSvar" group and "Utfall Svar" type group list. This allows for more dynamic control over the rendering of these elements.