Adds begrunnelse and bestemmelse to list unit
Extends the custom dispensations list unit to include "begrunnelse" (reasoning) and "bestemmelse" (provision) information.

These additions enhance the detail and context provided within the dispensations list, enabling more informed decision-making.
The title parameter in the custom field data is also updated to use dataTitle when it is a child component.