Adds workflows to publish to npm and Github Packages
Sets up GitHub Actions workflows for publishing the package to both npm and GitHub Packages.

This change replaces the single workflow with two separate workflows for npm and GitHub Packages, allowing for more targeted configuration and control over the publishing process.
It also ensures that Yarn v4 is used for the builds and publishing.