Adds custom list component for tiltaksformaal
Adds a new custom list component to display "TiltaksFormaal" data.

This component handles the rendering and management of a list of
"TiltaksFormaal" items, including cases where "kodebeskrivelse" is "Annet"
and requires appending the "simpleBinding" value.

The component is configurable through props and utilizes resource values
for localization. It also handles empty states by displaying a
custom-field element instead of a custom-list element.