Refactors resource binding logic for Ettersending
Updates resource binding generation for the CustomGroupEttersending component.

This change improves the way resource bindings are handled, making them more flexible and easier to configure. It also ensures that default resource keys are used if specific bindings are not provided, and ensures 'emptyFieldText' is correctly included/excluded based on 'hideIfEmpty' property.